Moving the capnp stuff out of the public headers!
