+++ include/circt/Dialect/ESI/cosim/Endpoint.h
@@ -7,7 +7,6 @@
 #ifndef CIRCT_DIALECT_ESI_COSIM_ENDPOINT_H
 #define CIRCT_DIALECT_ESI_COSIM_ENDPOINT_H
 
-#include "circt/Dialect/ESI/cosim/CosimDpi.capnp.h"
 #include <functional>
 #include <map>
 #include <memory>
